feat:ログをベースキャンバスの要素とできるように処理を抜き出した
