+++ Assets/_CryStar/Runtime/Battle/UI/BattleLogContents.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b8b70691296569642acfe98cb244ba95
